pyproject.toml
[project]
name = "yt-trending"
version = "0.2.0"
description = "YouTube 인기 급상승 트렌드 분석 대시보드"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "openai>=2.31.0",
    "google-generativeai>=0.8.6",
    "google-genai>=1.72.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.scripts]
yt-trending = "src.cli:main"
yt-server   = "backend.main:start"

[tool.hatch.build.targets.wheel]
packages = ["src", "backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
